DROID-1248 App | Enhancement | Storage limits (#3199)
* DROID-1248 mw

* DROID-1246 Settings | Enhancement | Space usage and file limits (#3177)

* DROID-1246 space usage command

* DROID-1246 space usage response

* DROID-1246 space usage useCase

* DROID-1246 space storage

* DROID-1246 use case di

* DROID-1246 filelimit core model

* DROID-1246 build provider

* DROID-1246 subscribeToSpace

* DROID-1246 file limits events

* DROID-1246 screen state

* DROID-1246 rename

* DROID-1246 pr fix

* DROID-1246 jobs added

* DROID-1246 name fix

* DROID-1247 App | Enhancement | Sync status filter (#3179)

* DROID-1247 fileSyncStatus

* DROID-1247 sync status relation

* DROID-1247 search file objects by sync status

* DROID-1248 Editor | Enhancement | Storage limits toast (#3181)

* DROID-1248 file storage, unsubscribe

* DROID-1248 show toast on file storage limit

* DROID-1248 tests

* DROID-1248 code style

* DROID-1248 fix

* DROID-1248 use case naming

* DROID-1248 rename const

* DROID-1248 jobs cancel

* DROID-1248 rename

* DROID-1248 tests

* DROID-1248 mw version

* DROID-1248 revert version

+++ core-utils/src/main/java/com/anytypeio/anytype/core_utils/ext/AndroidExtension.kt
@@ -395,7 +395,7 @@ fun getDeviceName(): String {
     }
 }
 
-fun bytesToHumanReadableSize(bytes: Double): String = when {
+fun bytesToHumanReadableSize(bytes: Long): String = when {
     bytes >= 1 shl 30 -> "%d GB".format((bytes / (1 shl 30)).toInt())
     bytes >= 1 shl 20 -> "%d MB".format((bytes / (1 shl 20)).toInt())
     bytes >= 1 shl 10 -> "%d kB".format((bytes / (1 shl 10)).toInt())
